chore:rename package name adeyawallet to id.credeble.adeya
