+++ android/app/build.gradle
@@ -75,9 +75,9 @@ android {
 
     compileSdkVersion rootProject.ext.compileSdkVersion
 
-    namespace "com.adeyawallet"
+    namespace "id.credebl.adeya"
     defaultConfig {
-        applicationId "com.adeyawallet"
+        applicationId "id.credebl.adeya"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
